+++ CHANGELOG.md
@@ -1,7 +1,23 @@
 # Changelog
 
-## 0.1.1
-- Bug fix: Reversing file content
+## [1.0.0] - 2025-02-25
+### Added
+- Windows installer
+- Logo changed
+- Unsaved changes require confirmation to be discarded
+- Shortcut: Use `CTRL+S` to save files
+- Shortcut: Use `CTRL+SHIFT+S` to save as...
+- Shortcut: Use `CTRL+O` to open files
+- Shortcut: Use `CTRL+N` to make new files
 
-## 0.1.0
+### Fixed
+- File extension bug while overwriting file
+- Permission error while saving generated user files, like custom settings
+
+## [0.1.1] - 2025-02-13
+### Fixed
+- Reversing file content
+
+## [0.1.0] - 2025-02-13
+### Added
 - First release